Split structs into files
